added docs for mjolnir
note that we will have to change how we use moveRotMotor because rn we assume that the user knows what the encoder counts. we assume that they don't, and keep it something like degrees / arcminutes / arcseconds / milliradians or anything more sensible that their stupid encoder counts, in order for ```mjolnir.py``` to be internally consistent.